refactor: unify login and signup page
In the world of OAuth and Magic Links, both the act of signing up and
logging in involves identical APIs. But we will still visually
distinguish them in the UI for the sake of user familiarity.

Note how things like names, usernames, etc. are not a part of the
initial registration process. We will redirect the user to a form to
complete this information if they are still missing from the database
(implement this later).
