+++ app/flow/signup/page.tsx
@@ -1,130 +1,8 @@
 "use client";
 
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { signIn } from "next-auth/react";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
-import React, { useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
-import { FaDiscord } from "react-icons/fa";
-import { FcGoogle } from "react-icons/fc";
-import { MdOutlineFacebook } from "react-icons/md";
-import { z } from "zod";
-
-const FormSchema = z.object({
-  email: z
-    .string()
-    .min(1, "E-mail address is required.")
-    .email("Not a valid e-mail address."),
-});
+import { AuthForm, AuthVariant } from "@/components/auth-form";
+import React from "react";
 
 export default function Page() {
-  const form = useForm<z.infer<typeof FormSchema>>({
-    resolver: zodResolver(FormSchema),
-  });
-
-  const [_errorMessage, setErrorMessage] = useState<string | null>(null);
-  const [loading, setLoading] = useState(false);
-  const router = useRouter();
-
-  const onSubmit = async (data: z.infer<typeof FormSchema>) => {
-    await signIn("resend", data);
-  };
-
-  useEffect(() => {
-    if (typeof window !== "undefined") {
-      if (localStorage.getItem("token")) {
-        router.push("/home");
-      }
-    }
-  }, [router]);
-
-  return (
-    <div className="px-7">
-      <div className="px-9">
-        <h1 className="font-bold text-3xl mt-auto">Sign up</h1>
-        <p className="text-zinc-400 mt-5 mb-12">
-          Come and build the next social phenomenon with us.
-        </p>
-
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
-            <FormField
-              control={form.control}
-              name="email"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>E-mail</FormLabel>
-                  <FormControl>
-                    <Input placeholder="john.smith@semicolon.com" {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <Button type="submit" className="w-full" loading={loading}>
-              <span className="ml-2 font-2xl">Sign up</span>
-            </Button>
-          </form>
-        </Form>
-
-        <p className="text-zinc-400 my-5 text-center">
-          Already have an account?{" "}
-          <Link
-            href="/flow/login"
-            className="text-sky-400 font-bold cursor-pointer hover:underline"
-          >
-            Sign in
-          </Link>
-        </p>
-
-        <div className="flex items-center gap-2 my-5">
-          <div className="w-full h-1 border-b-2"></div>
-          <span className="font-bold">Or</span>
-          <div className="w-full h-1 border-b-2"></div>
-        </div>
-
-        <div className="my-2">
-          <Button
-            variant="outline"
-            className="w-full flex gap-3 items-center text-zinc-700 dark:text-zinc-300 "
-            onClick={() => console.log("Sign up with Google")}
-          >
-            <FcGoogle className="size-6" />
-            <span>Sign up with Google</span>
-          </Button>
-        </div>
-        <div className="my-2">
-          <Button
-            variant="outline"
-            className="w-full flex gap-3 items-center text-zinc-700 dark:text-zinc-300"
-            onClick={() => console.log("Sign up with Facebook")}
-          >
-            <MdOutlineFacebook className="size-7 fill-[#2aa4f4]" />
-            <span>Sign up with Facebook</span>
-          </Button>
-        </div>
-        <div className="my-2">
-          <Button
-            variant="outline"
-            className="w-full flex gap-3 items-center text-zinc-700 dark:text-zinc-300 "
-            onClick={() => console.log("Sign up with Discord")}
-          >
-            <FaDiscord className="size-6 fill-[#536dfe]" />
-            <span>Sign up with Discord</span>
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
+  return <AuthForm variant={AuthVariant.SignUp} />;
 }
